+++ Idler/ShiftNote.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Idler.Interfaces;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlTypes;
